(chore) renamed inpatient to clinical-encounter view to match intent
