+++ packages/esm-patient-clinical-view-app/src/clinical-encounter/dashboard/in-patient.scss
