+++ packages/esm-patient-clinical-view-app/src/clinical-encounter/summary/out-patient-summary/out-patient-summary.scss
