+++ packages/esm-patient-clinical-view-app/src/clinical-encounter/summary/surgical summary/surgical-summary.component.tsx
@@ -9,7 +9,7 @@ import { ConfigObject } from '../../../config-schema';
 import { useSurgicalSummery } from '../../../hooks/useSurgicalSummery';
 import SummaryCard from '../summary-card.component';
 
-import styles from '../../in-patient-department/in-patient.scss';
+import styles from '../../dashboard/in-patient.scss';
 
 interface SurgicalSummeryProps {
   patientUuid: string;
