+++ packages/esm-patient-clinical-view-app/translations/en.json
@@ -22,6 +22,7 @@
   "clinicalEncounter": "Clinical Encounter",
   "clinicalViews": "Clinical views",
   "contacted": "Contacted",
+  "customViews": "In this section, you'll find custom clinical views tailored to patients' conditions and enrolled care programs.",
   "dateOfDelivery": "Date of Delivery",
   "dateOfSurgery": "Date of Surgery",
   "dead": "Dead",
@@ -55,7 +56,6 @@
   "htsResult": "Final result",
   "htsScreening": "HTS Screening",
   "htsTestType": "Test type",
-  "inPatient": "In-Patient",
   "inPatientSummary": "In Patient Summary",
   "labourAndDelivery": "Labour and Delivery",
   "maternalCondition": "Maternal Condition",
@@ -73,7 +73,6 @@
   "noObservationsFound": "No observations found",
   "operatingDoctor": "Operating Doctor",
   "otherSubstanceAbuse": "Other Substance Use",
-  "outPatient": "Out Patient",
   "partnerStatus": "HIV status of partner)",
   "placeOfDelivery": "Place of Delivery",
   "populationType": "Population type",
